+++ Runtime/Scripts/Logging/LogMessages.cs
@@ -242,6 +242,10 @@ namespace GLTFast.Logging
         /// Inconsistent use of vertex color attribute.
         /// </summary>
         InconsistentVertexColorUsage,
+        /// <summary>
+        /// Stream read error.
+        /// </summary>
+        StreamError,
     }
 
     /// <summary>
@@ -299,6 +303,7 @@ is approximated. Enable Opaque Texture access in Universal Render Pipeline!" },
             { LogCode.ShaderMissing, $"Shader \"{{0}}\" is missing. Make sure to include it in the build (see {GltfGlobals.GltfPackageName}/Documentation~/ProjectSetup.md#materials-and-shader-variants )" },
             { LogCode.SkinMissing, "Skin missing" },
             { LogCode.SparseAccessor, "Sparse Accessor not supported ({0})" },
+            { LogCode.StreamError, "Stream error: {0}." },
             { LogCode.TextureDownloadFailed, "Download texture {1} failed: {0}" },
             { LogCode.TextureInvalidType, "Invalid {0} texture type (material: {1})" },
             { LogCode.TextureLoadFailed, "Texture #{0} not loaded" },